Moved 'create_config' function from setup to runtime
When you now run 'alpsh' it checks to see if the history file exists, if
not it creates it.
This was previously done in the setup. But due too permission error it
made more sense to be created when the application is ran.

Fixes #1

+++ alpsh/history.py
@@ -1,20 +1,30 @@
 import json
 import time
+import os
+
+location = os.path.expanduser('~') + "/.alpsh/"
 
 
 def create():
-    data = {'history': []}
-    with open('alpsh_history.json', 'w') as history:
-        json.dump(data, history, indent=4)
+    if not os.path.exists(location):
+        os.makedirs(location)
+
+    if not os.path.isfile(location + "alpsh_history.json"):
+        data = {'history': []}
+        with open(location + 'alpsh_history.json', 'w') as history:
+            json.dump(data, history, indent=4)
 
 
 def write(command):
-    with open('alpsh_history.json') as history_read:
-        data = json.load(history_read)
-        filteredcommand = command.replace("\n", "")
-    data['history'].append({
-        'timestamp': time.strftime("%c"),
-        'command': filteredcommand
-    })
-    with open('alpsh_history.json', 'w') as history:
-        json.dump(data, history, indent=4)
+    try:
+        with open(location + 'alpsh_history.json') as history_read:
+            data = json.load(history_read)
+            filteredcommand = command.replace("\n", "")
+        data['history'].append({
+            'timestamp': time.strftime("%c"),
+            'command': filteredcommand
+        })
+        with open(location + 'alpsh_history.json', 'w') as history:
+            json.dump(data, history, indent=4)
+    except(IOError, OSError) as e:
+        print("File not found? : " + str(e))
